+++ src/main/java/com/infamous/dungeons_mobs/interfaces/IAquaticMob.java
@@ -19,8 +19,8 @@ public interface IAquaticMob {
         return blockPos.getY() < world.getSeaLevel() - 5;
     }
 
-    default <T extends LivingEntity & IAquaticMob> boolean okTarget(T aquaticMob, @Nullable LivingEntity target){
-        if(aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
+    default <T extends LivingEntity & IAquaticMob> boolean okTarget(T aquaticMob, @Nullable LivingEntity target) {
+        if (aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
 
         if (target != null) {
             return !aquaticMob.level.isDay() || target.isInWater();
@@ -29,8 +29,8 @@ public interface IAquaticMob {
         }
     }
 
-    default <T extends Mob & IAquaticMob> boolean wantsToSwim(T aquaticMob){
-        if(aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
+    default <T extends Mob & IAquaticMob> boolean wantsToSwim(T aquaticMob) {
+        if (aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
 
         if (this.isSearchingForLand()) {
             return true;
@@ -43,19 +43,19 @@ public interface IAquaticMob {
     void setSearchingForLand(boolean searchingForLand);
 
     default <T extends Mob & IAquaticMob> boolean closeToNextPos(T aquaticMob) {
-        if(aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
+        if (aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
 
         Path path = aquaticMob.getNavigation().getPath();
         if (path != null) {
             BlockPos blockpos = path.getTarget();
-            double distanceToSqr = aquaticMob.distanceToSqr((double)blockpos.getX(), (double)blockpos.getY(), (double)blockpos.getZ());
+            double distanceToSqr = aquaticMob.distanceToSqr(blockpos.getX(), blockpos.getY(), blockpos.getZ());
             return distanceToSqr < 4.0D;
         }
         return false;
     }
 
     default <T extends Mob & IAquaticMob> void updateNavigation(T aquaticMob) {
-        if(aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
+        if (aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
 
         if (!aquaticMob.level.isClientSide) {
             if (aquaticMob.isEffectiveAi() && aquaticMob.isInWater() && this.wantsToSwim(aquaticMob)) {
@@ -69,7 +69,7 @@ public interface IAquaticMob {
     }
 
     default <T extends Mob & IAquaticMob> void checkAquaticTravel(T aquaticMob, Vec3 travelVec) {
-        if(aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
+        if (aquaticMob != this) throw new IllegalArgumentException("Supplied aquaticMob is not this instance!");
 
         if (aquaticMob.isEffectiveAi() && aquaticMob.isInWater() && this.wantsToSwim(aquaticMob)) {
             aquaticMob.moveRelative(0.01F, travelVec);
